restructure deployment related CLI code
