+++ bentoml/cli/__init__.py
@@ -26,19 +26,7 @@ from bentoml.server.gunicorn_server import (
     get_gunicorn_worker_count,
 )
 from bentoml.cli.click_utils import DefaultCommandGroup, conditional_argument
-from bentoml.deployment.serverless import ServerlessDeployment
-from bentoml.deployment.sagemaker import SagemakerDeployment
-from bentoml.exceptions import BentoMLException
-
-SERVERLESS_PLATFORMS = ["aws-lambda", "aws-lambda-py2", "gcp-function"]
-
-CLICK_COLOR_SUCCESS = "green"
-CLICK_COLOR_ERROR = "red"
-
-
-def _echo(message, color=CLICK_COLOR_SUCCESS):
-    click.echo("BentoML: ", nl=False)
-    click.secho(message, fg=color)
+from bentoml.cli.deployment import config_deployment_commands
 
 
 def create_bento_service_cli(archive_path=None):
@@ -163,173 +151,7 @@ def create_bentoml_cli():
     # Commands created here aren't mean to be used from generated service archive. They
     # are used as part of BentoML cli commands only.
 
-    # pylint: disable=unused-variable
-
-    # Example usage: bentoml deploy /ARCHIVE_PATH --platform=aws-lambda
-    @_cli.command(
-        help="Deploy BentoML archive as REST endpoint to cloud services",
-        short_help="Deploy Bento archive",
-    )
-    @click.argument("archive-path", type=click.STRING)
-    @click.option(
-        "--platform",
-        type=click.Choice(
-            [
-                "aws-lambda",
-                "aws-lambda-py2",
-                "gcp-function",
-                "aws-sagemaker",
-                "azure-ml",
-                "algorithmia",
-            ]
-        ),
-        required=True,
-        help="Target platform that Bento archive is going to deployed to",
-    )
-    @click.option(
-        "--region",
-        type=click.STRING,
-        help="Target region inside the cloud provider that will be deployed to",
-    )
-    @click.option("--stage", type=click.STRING)
-    @click.option(
-        "--api-name", type=click.STRING, help="The name of API will be deployed"
-    )
-    @click.option(
-        "--instance-type",
-        type=click.STRING,
-        help="SageMaker deployment ONLY. The instance type will be used for deployment",
-    )
-    @click.option(
-        "--instance-count",
-        type=click.INT,
-        help="Sagemaker deployment ONLY. Number of instances will be used for \
-            deployment",
-    )
-    def deploy(
-        archive_path, platform, region, stage, api_name, instance_type, instance_count
-    ):
-        if platform in SERVERLESS_PLATFORMS:
-            deployment = ServerlessDeployment(archive_path, platform, region, stage)
-        elif platform == "aws-sagemaker":
-            deployment = SagemakerDeployment(
-                archive_path, api_name, region, instance_count, instance_type
-            )
-        else:
-            _echo(
-                'Deploying with "--platform=%s" is not supported in current version of '
-                "BentoML".format(platform),
-                CLICK_COLOR_ERROR,
-            )
-            return
-
-        output_path = deployment.deploy()
-
-        _echo("Successfully deployed to {platform}!".format(platform=platform))
-        _echo("Deployment archive is saved at: %s" % output_path)
-
-    # Example usage: bentoml delete-deployment ARCHIVE_PATH --platform=aws-lambda
-    @_cli.command(
-        help="Delete active BentoML deployment from cloud services",
-        short_help="Delete active BentoML deployment",
-    )
-    @click.argument("archive-path", type=click.STRING)
-    @click.option(
-        "--platform",
-        type=click.Choice(
-            [
-                "aws-lambda",
-                "aws-lambda-py2",
-                "gcp-function",
-                "aws-sagemaker",
-                "azure-ml",
-                "algorithmia",
-            ]
-        ),
-        required=True,
-        help="The platform bento archive is deployed to",
-    )
-    @click.option(
-        "--region",
-        type=click.STRING,
-        required=True,
-        help="The region deployment belongs to",
-    )
-    @click.option(
-        "--api-name",
-        type=click.STRING,
-        help="Name of the API function that is deployed",
-    )
-    @click.option("--stage", type=click.STRING)
-    def delete_deployment(archive_path, platform, region, stage, api_name):
-        if platform in SERVERLESS_PLATFORMS:
-            deployment = ServerlessDeployment(archive_path, platform, region, stage)
-        elif platform == "aws-sagemaker":
-            deployment = SagemakerDeployment(archive_path, api_name, region)
-        else:
-            _echo(
-                "Remove deployment with --platform=%s is not supported in current "
-                "version of BentoML".format(platform),
-                CLICK_COLOR_ERROR,
-            )
-            return
-
-        if deployment.delete():
-            _echo("Successfully delete {platform} deployment".format(platform=platform))
-        else:
-            _echo(
-                "Delete {platform} deployment unsuccessful".format(platform=platform),
-                CLICK_COLOR_ERROR,
-            )
-
-    # Example usage: bentoml check-deployment-status ARCHIVE_PATH --platform=aws-lambda
-    @_cli.command(
-        help="Check deployment status of BentoML archive",
-        short_help="check deployment status",
-    )
-    @click.argument("archive-path", type=click.STRING)
-    @click.option(
-        "--platform",
-        type=click.Choice(
-            [
-                "aws-lambda",
-                "aws-lambda-py2",
-                "gcp-function",
-                "aws-sagemaker",
-                "azure-ml",
-                "algorithmia",
-            ]
-        ),
-        required=True,
-        help="Target platform that Bento archive will be deployed to as a REST api \
-            service",
-    )
-    @click.option(
-        "--region",
-        type=click.STRING,
-        required=True,
-        help="Deployment's region name inside cloud provider.",
-    )
-    @click.option("--stage", type=click.STRING)
-    @click.option(
-        "--api-name",
-        type=click.STRING,
-        help="The name of API that is deployed as a service.",
-    )
-    def check_deployment_status(archive_path, platform, region, stage, api_name):
-        if platform in SERVERLESS_PLATFORMS:
-            deployment = ServerlessDeployment(archive_path, platform, region, stage)
-        elif platform == "aws-sagemaker":
-            deployment = SagemakerDeployment(archive_path, api_name, region)
-        else:
-            _echo(
-                "check deployment status with --platform=%s is not supported in the "
-                "current version of BentoML" % platform,
-                CLICK_COLOR_ERROR,
-            )
-            return
-
-        deployment.check_status()
+    config_deployment_commands(_cli)
 
     # pylint: enable=unused-variable
     return _cli
